refactor(reservation-domain): Waitlist 패키지로 도메인 분리
- Waitlist, 관련 Embedded, Enum 대상으로 domain.waitlist 패키지로 이동함
- Test, TestFixture 도 패키지 맞춤
