+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/waitlist/model/WaitlistNotification.java
@@ -1,6 +1,6 @@
 package com.seatliberator.seatliberator.reservation.application.waitlist.model;
 
-import com.seatliberator.seatliberator.reservation.domain.persistence.Waitlist;
+import com.seatliberator.seatliberator.reservation.domain.waitlist.Waitlist;
 
 public record WaitlistNotification(
         Waitlist request,
