+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/waitlist/port/in/CreateWaitlistUseCase.java
@@ -1,7 +1,7 @@
 package com.seatliberator.seatliberator.reservation.application.waitlist.port.in;
 
 import com.seatliberator.seatliberator.reservation.application.waitlist.port.in.command.CreateWaitlistCommand;
-import com.seatliberator.seatliberator.reservation.domain.persistence.Waitlist;
+import com.seatliberator.seatliberator.reservation.domain.waitlist.Waitlist;
 
 public interface CreateWaitlistUseCase {
     Waitlist create(CreateWaitlistCommand command);
