+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/waitlist/application/model/WaitlistRequestsTest.java
@@ -3,9 +3,9 @@ package com.seatliberator.seatliberator.reservation.waitlist.application.model;
 import com.seatliberator.seatliberator.reservation.application.waitlist.internal.WaitlistPromotionResult;
 import com.seatliberator.seatliberator.reservation.application.waitlist.model.WaitlistNotification;
 import com.seatliberator.seatliberator.reservation.application.waitlist.model.WaitlistRequests;
-import com.seatliberator.seatliberator.reservation.domain.*;
-import com.seatliberator.seatliberator.reservation.domain.fixture.WaitlistFixtureBuilder;
-import com.seatliberator.seatliberator.reservation.domain.persistence.Waitlist;
+import com.seatliberator.seatliberator.reservation.domain.SeatLocator;
+import com.seatliberator.seatliberator.reservation.domain.TimeRange;
+import com.seatliberator.seatliberator.reservation.domain.waitlist.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
