+++ reservation/reservation-domain/src/main/java/com/seatliberator/seatliberator/reservation/domain/waitlist/WaitlistStatus.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.reservation.domain;
+package com.seatliberator.seatliberator.reservation.domain.waitlist;
 
 public enum WaitlistStatus {
     ACTIVE,
