+++ reservation/reservation-persistence/src/main/java/com/seatliberator/seatliberator/reservation/persistence/waitlist/jpa/JpaWaitlistStore.java
@@ -3,8 +3,8 @@ package com.seatliberator.seatliberator.reservation.persistence.waitlist.jpa;
 import com.seatliberator.seatliberator.reservation.application.waitlist.port.out.WaitlistStore;
 import com.seatliberator.seatliberator.reservation.domain.SeatLocator;
 import com.seatliberator.seatliberator.reservation.domain.TimeRange;
-import com.seatliberator.seatliberator.reservation.domain.WaitlistStatus;
-import com.seatliberator.seatliberator.reservation.domain.persistence.Waitlist;
+import com.seatliberator.seatliberator.reservation.domain.waitlist.Waitlist;
+import com.seatliberator.seatliberator.reservation.domain.waitlist.WaitlistStatus;
 import com.seatliberator.seatliberator.reservation.persistence.shared.jpa.specification.CommonPredicates;
 import com.seatliberator.seatliberator.reservation.persistence.shared.jpa.specification.SeatLocatorPredicates;
 import com.seatliberator.seatliberator.reservation.persistence.shared.jpa.specification.TimeRangePredicates;
