+++ reservation/reservation-persistence/src/main/java/com/seatliberator/seatliberator/reservation/persistence/waitlist/jpa/repository/WaitlistRepository.java
@@ -1,6 +1,6 @@
 package com.seatliberator.seatliberator.reservation.persistence.waitlist.jpa.repository;
 
-import com.seatliberator.seatliberator.reservation.domain.persistence.Waitlist;
+import com.seatliberator.seatliberator.reservation.domain.waitlist.Waitlist;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
